Fix 32-bit Huffman size reads on LP64 platforms
On macOS (and other LP64 platforms) long is 64 bits. The original code
read the 4-byte expanded size with *(long *), pulling in 8 bytes and
creating a garbage size. Replace with int32_t in 3 locations:
- CA_Startup pictable load
- CAL_ExpandGrChunk size header
- CA_CacheScreen size header

+++ src/id_ca.c
@@ -313,7 +313,7 @@ void CA_Startup(void) {
             byte *source = malloc(compressed);
             lseek(grhandle, pos, SEEK_SET);
             read(grhandle, source, compressed);
-            long expanded = *(long *)source;
+            int32_t expanded = *(int32_t *)source;
             MM_GetPtr((memptr *)&pictable, expanded);
             CAL_HuffExpand(source + 4, (byte *)pictable, expanded, grhuffman);
             free(source);
@@ -389,7 +389,7 @@ void CAL_ExpandGrChunk(int chunk, byte *source) {
             expanded = 128 * 16;
     } else {
         // Everything else has an explicit size longword
-        expanded = *(long *)source;
+        expanded = *(int32_t *)source;
         source += 4;
     }
     
@@ -537,7 +537,7 @@ void CA_CacheScreen(int chunk) {
     source = malloc(compressed);
     read(grhandle, source, compressed);
     
-    expanded = *(long *)source;
+    expanded = *(int32_t *)source;
     dest = malloc(expanded);
     CAL_HuffExpand(source + 4, dest, expanded, grhuffman);
